+++ lizmap/forms/layout_edition.py
@@ -1,9 +1,10 @@
 """Dialog for layout edition."""
-
+from qgis.core import QgsMasterLayoutInterface, QgsProject
 from qgis.gui import QgsFileWidget
+from qgis.PyQt.QtCore import Qt
 from qgis.PyQt.QtGui import QIcon
 
-from lizmap.definitions.layouts import LayoutsDefinitions
+from lizmap.definitions.layouts import Dpi, FormatType, LayoutsDefinitions
 from lizmap.forms.base_edition_dialog import BaseEditionDialog
 from lizmap.qgis_plugin_tools.tools.i18n import tr
 from lizmap.qgis_plugin_tools.tools.resources import load_ui, resources_path
@@ -59,6 +60,54 @@ class LayoutEditionDialog(BaseEditionDialog, CLASS):
 
         self.setup_ui()
 
+    def post_load_form(self):
+        """ When the form has been loaded. """
+        manager = QgsProject.instance().layoutManager()
+        layout = manager.layoutByName(self.layout.text())
+
+        if layout.layoutType() == QgsMasterLayoutInterface.PrintLayout:
+            for _print_layout in manager.printLayouts():
+                if _print_layout.name() == self.layout.text():
+                    atlas_layout = _print_layout
+                    break
+
+        if not atlas_layout:
+            # The report shouldn't be loaded
+            return
+
+        is_atlas = atlas_layout.atlas().enabled()
+        self.icon.setEnabled(is_atlas)
+        self.label_icon.setEnabled(is_atlas)
+
+        self.formats.setEnabled(not is_atlas)
+        self.dpi.setEnabled(not is_atlas)
+        self.default_format.setEnabled(not is_atlas)
+        self.default_dpi.setEnabled(not is_atlas)
+        self.label_formats.setEnabled(not is_atlas)
+        self.label_dpi.setEnabled(not is_atlas)
+        self.label_default_format.setEnabled(not is_atlas)
+        self.label_default_dpi.setEnabled(not is_atlas)
+
+        # Reset some values
+        if is_atlas:
+            # Formats options
+            for index, item in enumerate(self.config.layer_config.get('formats_available')['items']):
+                self.formats.setItemCheckState(index, Qt.Checked if item == FormatType.Pdf else Qt.Unchecked)
+
+            # Default format
+            index = self.default_format.findData(FormatType.Pdf.value['data'])
+            self.default_format.setCurrentIndex(index)
+
+            # DPI options
+            for index, item in enumerate(self.config.layer_config.get('dpi_available')['items']):
+                self.dpi.setItemCheckState(index, Qt.Checked if item == Dpi.Dpi100 else Qt.Unchecked)
+
+            # Default DPI
+            index = self.default_dpi.findData(Dpi.Dpi100.value['data'])
+            self.default_dpi.setCurrentIndex(index)
+        else:
+            self.icon.setFilePath("")
+
     def validate(self) -> str:
         upstream = super().validate()
         if upstream:
@@ -74,5 +123,5 @@ class LayoutEditionDialog(BaseEditionDialog, CLASS):
 
     def open_wizard_group(self):
         """ Open the wizard about ACL. """
-        helper = tr("Setting groups for the layer editing capabilities '{}'".format(""))
+        helper = tr("Setting groups for the layout print capabilities'{}'".format(self.layout.text()))
         super().open_wizard_dialog(helper)